Movement update
Created a new movement script. It now has capped player-induced movement
speed, but allows you to move faster than that should an external force
cause you to move faster. It has a jump feature, but no detection
whether you are double jumping or on the ground, it just jumps whenever
you press the button. Changed the player model to a sphere, the square's
friction was preventing movement. Not sure how this is affected when we
have a player model, may need some recoding at that point.